Added default panels for each button displays its own page and information. I also added a new class.
The taskbar has been changed to the colour blue and the main background of each page has been changed to white.

I also added a test bit of text on each page that allows  me to see that each page runs on its own when the appropriate button in the taskbar is clicked.